# temp_thkiengiangshoes/thkiengiangshoes-main/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TBS II Production & Maintenance Management System API
  description: OpenAPI Specification for TBS II Multi-platform Factory Operations, SLA Calculation Engine, and BI Dashboard.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development API Server
paths:
  /auth/login:
    post:
      summary: User Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emp_code, password]
              properties:
                emp_code: { type: string, example: "CN001" }
                password: { type: string, example: "123456" }
      responses:
        '200':
          description: Successful authentication, returns JWT access token and user profile
...